Added get opertion for meal (yet to be tested)
also added tests to write all futre test
