+++ views/tests.py
@@ -1,0 +1,17 @@
+import requests
+
+
+# def test_meal():
+#     # Test adding a meal
+#     response = requests.post(f'{BASE_URL}/meal/', json={
+#         'name': 'pizza',
+#         'price': 45.0,
+#         'description': 'the best pizza in the world',
+#         'image': 'https://en.wikipedia.org/wiki/File:Pizza-3007395.jpg',
+#         'foodType': 0
+#     })
+#     print("Testing POST /meal/ endpoint:")
+#     print(f"Response: {response.json()}")
+#
+# if __name__ == '__main__':
+#     test_meal();
